[interfaces-config]: enable multi-asic support
* create service template
* enable template instances in sonic_debian_extension
* use iproute2 in interfaces-config script if working with network namespaces
* make swss multi-instance dependent on multi-instance interfaces config

+++ files/image_config/interfaces/interfaces-config.sh
@@ -1,11 +1,21 @@
 #!/bin/bash
 
-ifdown --force eth0
+NS=$1
 
-sonic-cfggen -d -t /usr/share/sonic/templates/interfaces.j2 > /etc/network/interfaces
+if [ -z "$(ip netns)" ]; then
+    ifdown --force eth0
+    sonic-cfggen -d -t /usr/share/sonic/templates/interfaces.j2 > /etc/network/interfaces 
+    [ -f /var/run/dhclient.eth0.pid ] && kill `cat /var/run/dhclient.eth0.pid` && rm -f /var/run/dhclient.eth0.pid
+    systemctl restart networking
+    ifdown lo && ifup lo
+else
+    INTF_LIST=`redis-cli $NS -n 4 KEYS 'LOOPBACK_INTERFACE*'`    
+    for ENTRY in $INTF_LIST; do
+        IFS='|' read -ra ADDR_ARRAY <<< "$ENTRY"
+        ADDR=${ADDR_ARRAY[2]}   
+        namespace $NS ip addr add $ADDR dev lo
+    done
+fi
 
-[ -f /var/run/dhclient.eth0.pid ] && kill `cat /var/run/dhclient.eth0.pid` && rm -f /var/run/dhclient.eth0.pid
 
-systemctl restart networking
 
-ifdown lo && ifup lo
